// client/src/constants.ts
export enum ActionTypes {
  LOGIN_SUCCESS = "LOGIN_SUCCESS",
  LOGIN_FAIL = "LOGIN_FAIL",

  LOGOUT_SUCCESS = "LOGOUT_SUCCESS",
  LOGOUT_FAIL = "LOGOUT_FAIL",

  UPDATE_USER = "UPDATE_USER",

  CREATE_USER_SUCCESS = "CREATE_USER_SUCCESS",
  CREATE_USER_FAIL = "CREATE_USER_FAIL",

  CREATE_TRANSACTION_SUCCESS = "CREATE_TRANSACTION_SUCCESS",
  CREATE_TRANSACTION_FAIL = "CREATE_TRANSACTION_FAIL",

  GET_TRANSACTIONS_FAIL = "GET_TRANSACTIONS_FAIL",
  GET_TRANSACTIONS_SUCCESS = "GET_TRANSACTIONS_SUCCESS",

  CLEAN_ERROR = "CLEAN_ERROR",
  IS_LOADING = "IS_LOADING",
}